+++ src/main/java/com/jeleniasty/betapp/features/user/dto/UserScoreDTO.java
@@ -1,3 +1,3 @@
 package com.jeleniasty.betapp.features.user.dto;
 
-public record UserScoreDTO(Long id, String username, int points) {}
+public record UserScoreDTO(Long id, String username, double points) {}
